==> 02_Categorization/01_Prototype_LM/01_preprocess_data.ipynb <==
## Description
In this notebook, the preprocessing of the German titles and descriptions of each entry is performed. The cleaned and combined text data will be used for keywordextraction in next step.

The code was created with the assistance of ChatGPT-4.

import pandas as pd
import re

inputdata_file = 'data/merged_dataset_metadata.csv'
outputdata_file ='data/01_preprocessed_data.csv'

df = pd.read_csv(inputdata_file, low_memory=False)

# Function for text preprocessing
def preprocess_text(text):

    # Remove HTML tags
    text = re.sub(r'<.*?>', '', text)
    # Remove URLs
    text = re.sub(r'http[s]?://\S+', '', text)
    # Remove file extensions
    text = re.sub(r'\b\w+\.(csv|json|shp|xls|parquet|rdfxml|jsonld|jsonl|dxf|gpkg|turtle)\b', '', text)
    # Remove dashes unless between letters
    text = re.sub(r'(?<!\w)-|-(?!\w)', '', text)
    # Replace symbols with space
    text = re.sub(r'\+[#|]', ' ', text)
    # Remove punctuation
    text = re.sub(r'[,«»‚’°%*;!?§±\'"/„“‘”\(\)\[\]]', '', text)
    text = re.sub(r'(?<=\s)[.:]|[.:](?=\s)', '', text)
    # Remove words with underscores
    text = re.sub(r'\b\w*_\w*\b', '', text)
    # Remove numbers
    text = re.sub(r'\b\d+\b', '', text)
    # Remove dashes unless between letters
    text = re.sub(r'(?<!\w)-|-(?!\w)', '', text)
    # Remove single characters
    text = re.sub(r'\b\w\b', '', text)
    # Replace multiple spaces with a single space
    text = re.sub(r'\s+', ' ', text)

    return text

# preprocess title, description and keyword DE:
df['dataset_title_DE_preprocessed'] = df['dataset_title_DE'].map(lambda x: preprocess_text(str(x)) if pd.notna(x) else x)
df['dataset_description_DE_preprocessed'] = df['dataset_description_DE'].map(lambda x: preprocess_text(str(x)) if pd.notna(x) else x)


# Write dataframe in new csv-File
df.to_csv(outputdata_file, index=False)

print(f'The file has been successfully saved as {outputdata_file}.')
